rename cpp_type_aux.py to cpp_dtype_aux.py
