+++ endf_parserpy/compiler/cpp_types/cpp_type_matrix2d/assign.py
@@ -34,7 +34,7 @@ from ..cpp_varaux import (
     get_cpp_varname,
     init_local_var_from_global_var,
 )
-from ..cpp_type_aux import map_dtype
+from ..cpp_dtype_aux import map_dtype
 from .query import Query
